fix migration errors

+++ database/migrations/2019_09_02_053155_create_payment_methods_table.php
@@ -13,13 +13,15 @@ class CreatePaymentMethodsTable extends Migration
      */
     public function up()
     {
-        Schema::create('payment_methods', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('name');
-            $table->integer('company_id')->unsigned()->nullable();
-            $table->foreign('company_id')->references('id')->on('companies')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('payment_methods')) {
+            Schema::create('payment_methods', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('name');
+                $table->integer('company_id')->unsigned()->nullable();
+                $table->foreign('company_id')->references('id')->on('companies')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
